requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
apprise>=1.8.0
PyYAML>=6.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
